refactor: Move PropertyTable into own file
